requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-whois==0.9.5
